move `filter` to its own file
